td-shim-ld: support use td-shim-ld as a library
Move the core code into lib.rs so it could be used as a library.
